eval/requirements.txt
torch==2.7.1
torchvision==0.22.1
torchaudio==2.7.1
triton==3.3.1
# diffusers==0.36.0
# transformers==4.57.6
# sentence-transformers==5.2.3
git+https://github.com/SHYuanBest/diffusers.git@test
git+https://github.com/huggingface/transformers.git
git+https://github.com/huggingface/sentence-transformers.git
accelerate==1.12.0
deepspeed==0.18.4
peft==0.18.1
huggingface-hub==1.4.1
zstandard==0.25.0
wandb==0.23.0
video-reader-rs==0.4.1
numpy<2.0.0
pandas
pillow
tqdm
scipy
opencv-python
scikit-image
ffmpeg-python
video-reader-rs
openai-clip
pyiqa
timm>=0.9
omegaconf
easydict
